Added validation logic and enums for error scenarios

+++ employeeserv-t23_24/employeeservApi/src/main/java/com/paypal/bfs/test/employeeserv/service/EmployeeService.java
@@ -7,5 +7,5 @@ public interface EmployeeService {
 	
 	public Employee createEmployee(Employee employee);
 	
-	public Optional<Employee> getEmployee(int id);
+	public Optional<Employee> getEmployee(long id);
 }
